% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.lp.R
\name{copy.lp}
\alias{copy.lp}
\title{copy.lp
This function is used to make a copy of a lpmodel that can be modified 
without changing the orginal lp}
\usage{
copy.lp(lps.model)
}
\arguments{
\item{lps.model}{the original lpmodel}
}
\value{
a pointer to a copy of the original lpmodel
}
\description{
copy.lp
This function is used to make a copy of a lpmodel that can be modified 
without changing the orginal lp
}
\examples{
lps.model <- lpSolveAPI::make.lp(4, 0)
res <- copy.lp(lps.model)
}
