% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineLPSolveMod.R
\name{defineLPSolveMod}
\alias{defineLPSolveMod}
\title{defineLPSolveMod
Specify a lpsolve odel (mainly for internal use)}
\usage{
defineLPSolveMod(
  A,
  b,
  C = NULL,
  v = NULL,
  lower = NULL,
  upper = NULL,
  maximum = TRUE,
  ob = NULL
)
}
\arguments{
\item{A}{a matrix of inequality constraints A x <=b}

\item{b}{a vector A x <=b}

\item{C}{a matrix of equality constraints C x = v, should be null in the
absence of such constraints}

\item{v}{a vector of equality constraints C x = v, should be null in the
absence of such constraints}

\item{lower}{minimal bounds for paramaters, by default set to zero}

\item{upper}{maximum bounds for paramaters, by default set to Inf}

\item{maximum}{tells whether the objective function is maximised or not}

\item{ob}{the coefficient of the ojective function (default all 1)}
}
\value{
returns an OP object
}
\description{
defineLPSolveMod
Specify a lpsolve odel (mainly for internal use)
}
