% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{degenerateSubSpace}
\alias{degenerateSubSpace}
\title{degenerateSubSpace
This function reduces the space defined by inequality and unequality 
constraints}
\usage{
degenerateSubSpace(A, b, C, v, z)
}
\arguments{
\item{A}{a matrix of coefficients of inequality constants A.x<=b}

\item{b}{a vector of length equals to nrow(A)}

\item{C}{a matrix of coefficients of inequality constants C.x=v}

\item{v}{a vector of length equals to nrow(C)}

\item{x0}{a vector of length equals to ncol(A) that should be in the polytope, for example returned by \code{\link{chebyCentre}}}
}
\value{
a list with elements A2 and b2 defining the subscape A2 x <= b2 and
Nt that can be used to convert results in appropriate format
}
\description{
degenerateSubSpace
This function reduces the space defined by inequality and unequality 
constraints
}
\examples{
n <- 5
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
C <- rbind(c(1,1,rep(0,n-2)),c(0,0,1,1,rep(0,n-4)))
v <- matrix(rep(0.2,2),2)
X0 <- rep(0.1,n)
x <- degenerateSubSpace(A,b,C,v,X0)
}
