% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSatiation.R
\name{estimateSatiation}
\alias{estimateSatiation}
\title{estimateSatiation
Estimate satiation based on allometric relationship}
\usage{
estimateSatiation(
  weight,
  taxon,
  assimilation_efficiency = 1,
  digestibility = 1
)
}
\arguments{
\item{weight}{wet weight in gram}

\item{taxon}{one of ("endotherm", #' "vertebrate ecotherm", "invertebrate")}

\item{assimilation_efficiency}{assimilation efficiency of the species
(unitless)}

\item{digestibility}{average digestibility correction factor of prey
(unitless)}
}
\value{
satiation as explained in supplementary material in
\insertCite{planque2022;textual}{RCaNmodel}
}
\description{
estimateSatiation
Estimate satiation based on allometric relationship
}
\details{
Following \insertCite{planque2022;textual}{RCaNmodel}, this function
estimates satiation based on an allometric relationship formulated by
\insertCite{yodzis1992;textual}{RCaNmodel}
}
\examples{
estimateSatiation(1e-4, taxon = "invertebrate", 1, 0.9)
}
\references{
\insertAllCited{}
}
