% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInitPoint.R
\name{findInitPoint}
\alias{findInitPoint}
\title{findInitPoint}
\usage{
findInitPoint(A, b, C = NULL, v = NULL)
}
\arguments{
\item{A}{a matrix of inequality constraints A x <=b}

\item{b}{a vector A x <=b}

\item{C}{a matrix of equality constraints C x = v, should be null in the
absence of such constraints}

\item{v}{a vector of equality constraints C x = v, should be null in the
absence of such constraints}
}
\value{
a random initial point or a vector of NA if enable to find a solution
}
\description{
find a random initial point by averaging results of random lp problems
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
findInitPoint(A = A, b = b)

}
