% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDerivedSymbolicObjects.R
\name{generateDerivedSymbolicObjects}
\alias{generateDerivedSymbolicObjects}
\title{generateDerivedSymbolicObjects}
\usage{
generateDerivedSymbolicObjects(
  param,
  env,
  before = FALSE,
  after = FALSE,
  delta = FALSE,
  beforedelta = FALSE,
  ratio = FALSE,
  beforeratio = FALSE
)
}
\arguments{
\item{param}{namesof param}

\item{env}{the environment to be updated}

\item{before}{whether to generate vector NA 1:T-1}

\item{after}{whether to generate vector 2:T NA}

\item{delta}{whether to generate Yt+1-Yt}

\item{beforedelta}{whether to generate Yt-Yt-1}

\item{ratio}{whether to generate Yt+1/Yt}

\item{beforeratio}{whether to generate Yt/Yt-1}
}
\value{
nothing but update the environment
}
\description{
This is an internal function that builds derived symbolic objects
required for the computation of the model
}
