% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSymbolicObjects.R
\name{generateSymbolicObjects}
\alias{generateSymbolicObjects}
\title{generateSymbolicObjects}
\usage{
generateSymbolicObjects(
  components,
  fluxes_def,
  ntstep,
  series,
  aliases,
  dynamics_equation = NULL,
  stanza_species,
  stanza,
  fluxes_stanza
)
}
\arguments{
\item{components}{the components data frame}

\item{fluxes_def}{fluxes_def}

\item{ntstep}{number of the time step}

\item{series}{the names of the series}

\item{aliases}{table of alias (default = NULL)}

\item{dynamics_equation}{a string that specifies the dynamics equation (NULL
if trophic model)}

\item{stanza_species}{a vector with the names of multistazas component}

\item{stanza}{a list with the parameters of the stanzas}

\item{fluxes_stanza}{a list with the flow names and the subflows for stanzas}
}
\value{
en environment storing all symbolic elements
}
\description{
This is an internal function that builds all the required symbolic objects
required for the computation of the model
}
