% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundParam.R
\name{getBoundParam}
\alias{getBoundParam}
\title{getBoundParam
Computes the possible bounds for a parameter p of a polytope defined
by A.x<=b and C.x=v or by the polytope of the CaNmod object}
\usage{
getBoundParam(x, p)
}
\arguments{
\item{x}{either a CaNmod oject or a named list with at least a matrix A and
a vector b (A.x<=b) and optionally a matrix C and a vector v (C.x=v)}

\item{p}{the index of the parameter for which the bounds should be computed}
}
\value{
a vector with lower bounds and upper bounds
}
\description{
getBoundParam
Computes the possible bounds for a parameter p of a polytope defined
by A.x<=b and C.x=v or by the polytope of the CaNmod object
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- getBoundParam(list(A = A, b = b),1)

}
