% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimerestrictedRCaNfile.R
\name{getTimerestrictedRCaNfile}
\alias{getTimerestrictedRCaNfile}
\title{getTimerestrictedRCaNfile}
\usage{
getTimerestrictedRCaNfile(path, outfile, from = NULL, to = NULL)
}
\arguments{
\item{outfile}{where to write the resulting RCaNfile}

\item{from}{first year of the restricted range, if NULL (default), the same
as the initial model year}

\item{to}{last year of the restricted range, if NULL (default), the same
as the last model year}

\item{file}{path to the orginal RCaN file}
}
\value{
nothing but writes the file where specified
}
\description{
This function generates a RCaN file restricted to a specific time range
based on another RCaN file
}
\examples{
infile <- system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaNmodel")
outfile <- paste0(tempfile(), ".xlsx")
getTimerestrictedRCaNfile(infile, outfile, from = 1989)
}
