% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggNetwork.R
\name{ggNetwork}
\alias{ggNetwork}
\title{ggNetwork
produces a ggraph representing the flow foodweb}
\usage{
ggNetwork(myCaNmod)
}
\arguments{
\item{myCaNmod}{a CaNmod oject}
}
\value{
a ggraph
}
\description{
ggNetwork
produces a ggraph representing the flow foodweb
}
\examples{
require(ggraph)
require(ggplot2)
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaNmodel"))
ggNetwork(myCaNmod)
}
