% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPairsBiomass.R
\name{ggPairsBiomass}
\alias{ggPairsBiomass}
\title{ggPairsBiomass
plots pairs plot of biomass among species and provided Kendall correlation
tau}
\usage{
ggPairsBiomass(
  mysampleCaNmod,
  species = NULL,
  years = NULL,
  frac = 1,
  logscale = TRUE,
  ...
)
}
\arguments{
\item{mysampleCaNmod}{result sent by \link{sampleCaN}}

\item{species}{the name (or a vector of name) of the species of interest
by default, all species}

\item{years}{years to be plotted (default all)}

\item{frac}{fraction of points to be plot (default all)}

\item{logscale}{should biomass be log10 transformed (default yes)}

\item{...}{other arguments sent to}
}
\value{
a ggplot
}
\description{
ggPairsBiomass
plots pairs plot of biomass among species and provided Kendall correlation
tau
}
\details{
distribution of fluxes or biomass for a specific year
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaNmodel"))
res <- sampleCaN(myCaNmod, 100)
ggPairsBiomass(res)

}
