% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mve_solver}
\alias{mve_solver}
\title{mve_solver
Find the maximum volume ellipsoid}
\usage{
mve_solver(A, b, x0, reg, x, E2, maxiter = 50L, tol = 1e-06)
}
\arguments{
\item{A}{a matrix of coefficients of inequality constants A.x<=b}

\item{b}{a vector of length equals to nrow(A)}

\item{x0}{a solution of the polytope the column scaling vector}

\item{reg}{a tuning parameter}

\item{x}{vector to store the center of the ellipse}

\item{E2}{matrix to store the ellipse E'E}

\item{maxiter}{a tuning parameter}

\item{tol}{a tuning parameter}
}
\description{
mve_solver
Find the maximum volume ellipsoid
}
\section{Details}{

The CHRR algorithm is a C++ translation of cobratoolbox code written
by Yin Zhang licensed under GNU GPL-3 https://github.com/opencobra/cobratoolbox/ 
and of matlab code written by Ben Cousins 
(https://github.com/Bounciness/Volume-and-Sampling) is a C++ translation of
matlab functions provided in the opencobra toolbox and 
in this github repository % `m x n` sparse matrix `A`.

Find the maximum volume ellipsoid
   {v:  v = x + Es, ||s|| <= 1}
 inscribing a full-dimensional polytope
         {v:  Av <= b}
 Input:  A, b --- defining the polytope
         x0 --- interior point (Ax0 < b)
 Output: x --- center of the ellipsoid
         E2 --- E'*E
--------------------------------------
Yin Zhang, Rice University, 07/29/02
--------------------------------------
lines modified by me (Ben Cousins) have a %Ben after them
}

