% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{round}
\alias{round}
\title{round
this function rounds the polytope and is adapted from the preprocess matlab
function}
\usage{
round(A, b, x0, N_total, p_shift, T, maxiter = 80L)
}
\arguments{
\item{A}{a matrix of coefficients of inequality constants A.x<=b}

\item{b}{a vector of length equals to nrow(A)}

\item{N_total}{the to store the transformation}

\item{p_shift}{the vector storing the total shift}

\item{T}{a matrix}

\item{maxiter}{a tuning parameter}
}
\description{
round
this function rounds the polytope and is adapted from the preprocess matlab
function
}
\section{Details}{

The CHRR algorithm is a C++ translation of cobratoolbox code written
by Yin Zhang licensed under GNU GPL-3 https://github.com/opencobra/cobratoolbox/ 
and of matlab code written by Ben Cousins 
(https://github.com/Bounciness/Volume-and-Sampling) is a C++ translation of
matlab functions provided in the opencobra toolbox and 
in this github repository % `m x n` sparse matrix `A`.

The algorithms rounds the polytope for
the volume/sampling algorithms to be accurate and efficient.
}

