% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCaN.R
\name{sampleCaN}
\alias{sampleCaN}
\title{sampleCaN}
\usage{
sampleCaN(
  myCaNmod,
  N,
  nchain = 1,
  ncore = 1,
  thin = 1,
  method = "gibbs",
  lastF = FALSE,
  keepCovMat = FALSE
)
}
\arguments{
\item{myCaNmod}{a CaNmod or a sampleCaNmod. In this latter case, adapative
phase and discarding phases won't be necessary}

\item{N}{the number of samples required}

\item{nchain}{the number of mcmc chains}

\item{ncore}{number of cores to use}

\item{thin}{thinning interval}

\item{method}{one of CDHR (default from volesti), RDHR (from volesti), gibbs}

\item{lastF}{should flow for last year be simulated (default = FALSE)}

\item{keepCovMat}{save or not the covariance matrix (saving can gain time
if new samples should be run but is very memory consuming - default FALSE)}
}
\value{
a sampleCaNmod object which contains three elements
\describe{
 \item{"CaNmod"}{the CaNmod object descring the model}
 \item{"mcmc"}{\code{\link[coda]{mcmc.list}}}
 \item{"x0"}{average starting point}
 \item{"covMat"}{average estimated covariance matrix that can be used
 for other run, if kept}
 \item{"N"}{reminder of the set up of the run}
 \item{"thin"}{reminder of the set up of the run}
 \item{"nchain"}{reminder of the set up of the run}
 \item{"method"}{reminder of the set up of the run}
}
}
\description{
sample the polytope corresponding to the CaNmod model
}
\examples{
myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx",
 package = "RCaNmodel"))
res <- sampleCaN(myCaNmod, 100)

}
