% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDerivedSymbolicObjects.R
\name{generateDerivedSymbolicObjects}
\alias{generateDerivedSymbolicObjects}
\title{generateDerivedSymbolicObjects}
\usage{
generateDerivedSymbolicObjects(
  param,
  env,
  M = FALSE,
  P = FALSE,
  delta = FALSE,
  deltaM = FALSE,
  ratio = FALSE,
  ratioM = FALSE
)
}
\arguments{
\item{param}{namesof param}

\item{env}{the environment to be updated}

\item{M}{whether to generate vector 2:T NA}

\item{P}{whether to generate vector NA 1:t-1}

\item{delta}{whether to generate Yt+1-Yt}

\item{deltaM}{whether to generate Yt-Yt-1}

\item{ratio}{whether to generate Yt+1/Yt}

\item{ratioM}{whether to generate Yt/Yt-1}
}
\value{
nothing but update the environment
}
\description{
This is an internal function that builds derived symbolic objects
required for the computation of the model
}
