% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSymbolicObjects.R
\name{generateSymbolicObjects}
\alias{generateSymbolicObjects}
\title{generateSymbolicObjects}
\usage{
generateSymbolicObjects(
  components,
  fluxes_def,
  ntstep,
  series,
  aliases,
  dynamics_equation = NULL
)
}
\arguments{
\item{components}{the components data frame}

\item{fluxes_def}{fluxes_def}

\item{ntstep}{number of the time step}

\item{series}{the names of the series}

\item{aliases}{table of alias (default = NULL)}

\item{dynamics_equation}{a string that specifies the dynamics equation (NULL
if trophic model)}
}
\value{
en environment storing all symbolic elements including among others
\itemize{
 \item{"IE_H"}{identity- H the matrix of the equation (Bt+1=(I-H).B+N.F)}
 \item{"n"}{the N matrix of the equation (Bt+1=(I-H).B+N.F)}
 \item{"B_0"}{initial biomasses}
 \item{"species"}{a vector of biomasse per time step for each species}
 \item{"flow"}{a vector of biomasse per time step for each flow}
}
}
\description{
This is an internal function that builds all the required symbolic objects
required for the computation of the model
}
